add migration and simple command handler for wallet creation
