+++ services/wallet/Wallet.Api/appsettings.json
@@ -1,4 +1,7 @@
 {
+  "ConnectionStrings":{
+    "EventStore":"Host=localhost;Port=5432;Database=WalletEvents;Username=admin;Password=admin"
+  },
   "Logging": {
     "LogLevel": {
       "Default": "Information",
