+++ services/wallet/Wallet.Application/ApplicationInstaller.cs
@@ -1,0 +1,13 @@
+﻿using Microsoft.Extensions.DependencyInjection;
+using Wallet.BuildingBlocks.Domain;
+using Wallet.Domain.WalletAggregate.Snapshots;
+
+namespace Wallet.Application;
+
+public static class ApplicationInstaller
+{
+    public static void ConfigureApplication(this IServiceCollection services)
+    {
+       
+    }
+}
