--- services/wallet/Wallet.Application/Class1.cs
@@ -1,5 +1,0 @@
-﻿namespace Wallet.Application;
-
-public class Class1
-{
-}
