+++ services/wallet/Wallet.Application/Commands/CreateWallet/CreateWalletCommand.cs
@@ -1,0 +1,11 @@
+﻿using MediatR;
+
+namespace Wallet.Application.Commands.CreateWallet;
+
+public class CreateWalletCommand : IRequest
+{
+    public decimal InitialBalance { get; set; }
+    public long OwnerUserId { get; set; }
+    public string OwnerMobile { get; set; }
+    public decimal OverUsedThreshold { get; set; }
+}
