+++ services/wallet/Wallet.Application/Wallet.Application.csproj
@@ -6,4 +6,16 @@
         <Nullable>enable</Nullable>
     </PropertyGroup>
 
+    <ItemGroup>
+      <Folder Include="Queries\" />
+    </ItemGroup>
+
+    <ItemGroup>
+      <PackageReference Include="MediatR" Version="12.4.1" />
+    </ItemGroup>
+
+    <ItemGroup>
+      <ProjectReference Include="..\Wallet.Infrastructure\Wallet.Infrastructure.csproj" />
+    </ItemGroup>
+
 </Project>
