--- services/wallet/Wallet.Infrastructure/Class1.cs
@@ -1,5 +1,0 @@
-﻿namespace Wallet.Infrastructure;
-
-public class Class1
-{
-}
